--- airflow/utils/env_versions.py
@@ -1,32 +1,0 @@
-# Licensed to the Apache Software Foundation (ASF) under one
-# or more contributor license agreements.  See the NOTICE file
-# distributed with this work for additional information
-# regarding copyright ownership.  The ASF licenses this file
-# to you under the Apache License, Version 2.0 (the
-# "License"); you may not use this file except in compliance
-# with the License.  You may obtain a copy of the License at
-#
-#   http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing,
-# software distributed under the License is distributed on an
-# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-# KIND, either express or implied.  See the License for the
-# specific language governing permissions and limitations
-# under the License.
-
-"""Utility functions to check the version of the environment packages and python version."""
-from __future__ import annotations
-
-from importlib.metadata import version
-
-from packaging.version import Version, parse as parse_version
-
-
-def _get_lib_major_version(lib_name: str) -> int:
-    ver: Version = parse_version(version(lib_name))
-    return ver.major
-
-
-def is_sqlalchemy_v1() -> bool:
-    return _get_lib_major_version("sqlalchemy") == 1
